Update doc impact to only make bugs on merge
Throughout the Havana release, the use of doc impact in commit
messages has created many helpful bugs for the documentation team.

However, the initial plan where the flags inclusion would alert doc
team to take a look at patches while under review has not ocurred,
due to a lack of people. In addition, abandoned patches have been
a distraction.

As such, the docs team wishes to change the behaviour of doc impact
so that is creates bugs once patches have been merged. This patch
simply changes the logic to realise this.

Change-Id: I29bab0095719212f6c5e2b9c6b704ed328976b2b