New headers argument for github.Issue.Issue
* jeepyb/cmd/close_pull_requests.py: The pygithub module added a new
positional argument for headers in the middle of other positional
arguments to the __init__ method for CompletableGithubObject. This
change passes an empty dict in for it to get the pull request closer
working again, and also adds names for the other arguments so they
will be easier to diagnose if this happens again.

Change-Id: Iea2913a9c76b564e9e5186fb970032696d292c86