Create local git mirror _after_ creating project
* jeepyb/cmd/manage_projects.py: Gerrit's create-project API calls
are refused with a "fatal: project xyzzy exists" error if its
repository is already present on disk. In fact, the same is true if
it's configured to perform local replication and the destination
replication directory for a new project already exists.
Conditionally create these directories only after creating the
project in Gerrit, avoiding further pain.

Change-Id: I3d7545a348187dfce3eba1ec93a976bb510668f3