Allow automatic subscription to DocImpact bugs
Our team's tech writer has a problem finding the DocImpact bugs
that our team has created, which is something she wants to do so
that she can make sure our team's work is well documented.

Instead of writing a horrible script to hack around this, I thought
I'd fix the problem generically, as I can imagine that this is
something that other teams might be interested in implementing.

I've also added some documentation on how to test this script, as
well as a sample config file for the subscriptions handling. If
someone wants to point me to the right place in puppet to add a
real configuration file, I am happy to do that too.

For reference, the configuration file is passed with the
--auto-subscriptions flag and the format of the configuration
file is documented in the code itself.

Change-Id: I5553bf94107912bc019cea4baedf639cadc6e314