Add script to post a gerrit welcome message
This script, to be triggered by the patchset-created hook in gerrit,
will post a welcoming message on new contributions from people who
have never contributed to OpenStack before.

The aim is to increase awareness of the place of code review in our
community, and encourage users to participate, while managing
expectations.

The plan is to merge this, then test thoroughly on review-dev with
a range of cases, including:
* First time patch (post message)
* Patch after change of email (no message)
* Patch after existing patch (no message)

Change-Id: I7bb2baeb37d3700ff07f39a8a29b4eaf41f5cd40
Partial-Bug: 1246879