Do github last
github project creation is really only a convenience, but API errors
cause other things to get borked from time to time. Move the interaction
with github to the very end, after we've done all of the things that
exist in our own backyard. Additionally, if we create the project in
github, it's possible we did it later on, so go ahead and trigger a
replication

Change-Id: I51572afe41f7ec9977ea7c17a90bd4df49b9a0f1
Closes-bug: #1242569