Add dryrun flag to welcome_message.py
In order to test this properly, we need a way to specify
not to post the message. This patch adds a new flag "--dryrun"
which will prevent the invocation of the method that posts
the message.

Minor logging changes were made to ensure we can still get info
about what's going on when in dry run mode.

Change-Id: I5cdb00bf89a1a83d70d03ec4a2ca7aae7651b4b3