Care less about github metadata
We actually make API calls to github on every run to make sure
that the description and homepage fields are set properly, as
well as wiki and downloads booleans... even though we've created
the last hundred accounts via manage-projects and it's 100% unlikely
that any of these have changed. We're also hitting github api limit
unhappiness with a full run.

Stop running all of these unless we just created the project.

Change-Id: I2ba958129efad80973bb5ba828a72210756c08de