Only add upstream remotes if we track upstream
Current runs fetch remote refs every time. This is because they were
not paying attention to the track-upstream config option in
the code where they add a git remote. This made the subsequent
git remote update command do way too much work.

Only have an upstream remote if we need one - and remove it if
we are configured to not need one.

Change-Id: Iff1bb2ddce3ec27d19fbbc246a5a3af26bceff29