Log project names as the projects are processed
Currently info-level logs only contain project names when errors
occur, and otherwise have a lot of information about branch processing
with no information about which projects the branches belong to. This
adds an info-level message with the project name so that well-behaved
runs of manage-projects have enough context to give meaningful logs.

Change-Id: Ibea0a8d8367da1b6e20d244e8117ddc241337e88