Refactor logging arguments into a common module
The command line utilities provided only haphazardly set up logging
using command line arguments. Refactor it into a common module, and
make use of it in all scripts that use logging.

This changes expire-old-reviews.py to not hardcode the log file, but
that is fine, it has no scripts calling it.

Change-Id: Ibc85f8e3b47f9c7898ad4334511b44e91ecbd736