Split track-upstream from manage-projects
We spend a very large amount of time in track-upstream code now because
of all of the debian repos. That's fine - but it makes iterating on
project creation issues exceptionally difficult. Also, now that we're
caching the flag as to whether or not the project is created,
track-upstream can read it and know whether or not it's safe to run for
a given project.

Change-Id: I454b1ba400dc86abcc9b939564eb4eb7c324308c